+++ src/components/content/mdx-components.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 
 import { Callout } from '@/components/callout';
-import { LinkPreview } from '@/components/link-preview';
+import { LinkPreview } from '@/components/content/link-preview';
 import { MdxCard } from '@/components/mdx-card';
 import { cn } from '@/lib/utils';
 import { MDXComponents } from 'mdx/types';
